MAYA-129276 move code to another location
