+++ lib/mayaUsd/resources/scripts/mayaUsdUtils.py
@@ -169,4 +169,28 @@ def setUserSelectedUSDDialogFileFilter(fileFilter):
     _userSelectedUSDDialogFileFilter = fileFilter
     
     
+def wantReferenceCompositionArc():
+    opVarName = "mayaUsd_WantReferenceCompositionArc"
+    return cmds.optionVar(exists=opVarName) and cmds.optionVar(query=opVarName)
+
+def saveWantReferenceCompositionArc(want):
+    opVarName = "mayaUsd_WantReferenceCompositionArc"
+    cmds.optionVar(iv=(opVarName, want))
+
+def wantPrependCompositionArc():
+    opVarName = "mayaUsd_WantPrependCompositionArc"
+    return cmds.optionVar(exists=opVarName) and cmds.optionVar(query=opVarName)
+
+def saveWantPrependCompositionArc(want):
+    opVarName = "mayaUsd_WantPrependCompositionArc"
+    cmds.optionVar(iv=(opVarName, want))
+
+def wantPayloadLoaded():
+    opVarName = "mayaUsd_WantPayloadLoaded"
+    return cmds.optionVar(exists=opVarName) and cmds.optionVar(query=opVarName)
+
+def saveWantPayloadLoaded(want):
+    opVarName = "mayaUsd_WantPayloadLoaded"
+    cmds.optionVar(iv=(opVarName, want))
+
